sum_adjacent_numbers works
I got the function to add the adjacent numbers to work in the given example, but there are some issues with formatting the output in the driver that I haven't gotten around to yet. For example, it still doesn't save the dimension and discard that info (it just goes through the sum_adjacent_numbers method like everything else in the file).